src/requirements.txt
apiclient
email_validator
flask
flask-wtf
google-api-python-client
google-auth
google_auth_oauthlib
httplib2
oauth2
pyyaml
uWSGI
werkzeug
wtforms
